# tools/openapi.yaml
openapi: 3.1.0
info:
  title: CofounderAgent Tool Service
  version: 0.4.0
  description: |
    OpenAPI surface for tools exposed to the CofounderAgent ReAct loop and
    optionally to external orchestrators (MCP, other agents). Mirrors the
    OpenAI-style function schemas in `src/server/toolsSchema.ts` and the
    Python `@tool` decorators in `src/CofounderAgent/main.py`.

    Authentication: local-only mode. Endpoints do not require login,
    sessions, bearer tokens, or CSRF headers. Ownership is derived from the
    server-side local principal and checked where a conversation/task id is
    provided.

    Approval gating: tools tagged `destructive` (`write_file`,
    `search_replace`, `run_command`, `generate_image`) create an `Approval`
    row when `toolsMode != "allowed"`. The loop polls until the user
    resolves the approval, then re-executes the tool.

    Rollback: every successful `write_file` / `search_replace` snapshots the
    pre-change content into `/tmp/cofounder_rollback_*` and writes a
    `RollbackSnapshot` DB row. Snapshots older than 14 days are sweepable via
    `DELETE /api/rollback`.
servers:
  - url: http://127.0.0.1:3000
    description: Local Next.js backend (loopback only by default).
...
paths:
...
  /api/tools/execute:
    post:
      operationId: executeTool
      summary: Generic tool dispatcher (proxies into the ReAct executor).
      description: |
        Single entrypoint that routes to read_file / list_directory /
        write_file / search_replace / run_command / grep / fetch_url /
        git_status / git_diff. Used by the chat loop. Direct callers should
        prefer the specific endpoints below for clearer audit trails.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ToolCall"
      responses:
        "200":
          description: ToolResult
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ToolResult" }

  # ── UI-facing low-risk tools (legacy v1 surface) ─────────────────────
...
  /api/files/search:
    get:
      operationId: search_files
      tags: [search]
      summary: Fuzzy file path search for the @-mention UI. Read-only.
      parameters:
        - name: q
          in: query
          schema: { type: string }
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 50, default: 20 }
      responses:
        "200":
          description: matches
          content:
            application/json:
              schema:
                type: object
                properties:
                  matches:
                    type: array
                    items: { type: string }
components:
  schemas:
    ToolCall:
      type: object
      properties:
        name:
          type: string
          enum:
            - read_file
            - list_directory
            - write_file
            - search_replace
            - run_command
            - grep
            - fetch_url
            - generate_image
            - git_status
            - git_diff
            - git_log
            - git_branch
            - git_show
            - add
            - multiply
        arguments:
          type: object
          additionalProperties: true
        call_id: { type: string }
      required: [name, arguments]
    ToolResult:
      type: object
      properties:
        call_id: { type: string }
        name: { type: string }
        output: { type: string }
        status:
          type: string
          enum: [ok, error, awaiting_approval]
        approvalId: { type: string, nullable: true }
        diff: { type: string, nullable: true }
        rollbackDir: { type: string, nullable: true }
      required: [name, output, status]